requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.0
stripe>=12.0.0
python-dotenv>=1.0.0
openai>=1.71.0